backend/pyproject.toml
[project]
name = "ai-content-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "alembic>=1.16.2",
    "beautifulsoup4>=4.13.4",
    "celery>=5.5.3",
    "fastapi>=0.115.13",
    "feedparser>=6.0.11",
    "langchain-openai>=0.3.25",
    "langgraph>=0.4.8",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.1.1",
    "PyYAML>=6.0",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "scrapy>=2.13.2",
    "sqlalchemy>=2.0.41",
    "tenacity>=8.2.0",
    "httpx>=0.24.0",
    "uvicorn>=0.34.3",
    "langchain>=0.3.26",
]
